refactor: read_adc 함수의 매개변수 타입을 uint8_t에서 int로 변경

+++ counters/rpi-counter-sensor/src/bpw34_sensor_data_reader.cpp
@@ -31,7 +31,7 @@ void Bpw34SensorDataReader::initialize_spi(const std::string& spi_dev_path) {
     }
 }
 
-uint16_t Bpw34SensorDataReader::read_adc(uint8_t channel) {
+uint16_t Bpw34SensorDataReader::read_adc(int channel) {
     if (channel < 0 || channel > 7) {
         throw std::invalid_argument("Invalid channel number");
     }
